+++ resources/views/components/sidebar/brand.blade.php
@@ -1,7 +1,0 @@
-@props([
-    'href' => '/',
-])
-
-<a href="{{ $href }}" class="aura-sidebar-brand" {{ $attributes }}>
-    {{ $slot }}
-</a>